DAT-3769 | extract infobox builder diff methods to util
