+++ front/main/app/models/infobox-builder.js
@@ -272,50 +272,6 @@ const InfoboxBuilderModel = Ember.Object.extend({
 	},
 
 	/**
-	 * @desc Provides information about whether and what elements'
-	 * values has been changed, basing on previously saved old data.
-	 *
-	 * @returns {Array} of Objects with element type and changed property
-	 */
-	createDataDiffs() {
-		const currentState = this.get('infoboxState'),
-			diffs = [];
-		let diff = [];
-
-		currentState
-			.filter((item) => item.infoboxBuilderData.originalData)
-			.forEach((item) => {
-				switch (item.type) {
-					case 'title':
-						diff = InfoboxBuilderModel.createTitleDiff(
-							item.infoboxBuilderData.originalData,
-							item.data
-						);
-						break;
-					case 'row':
-						diff = InfoboxBuilderModel.createRowDiff(
-							item.infoboxBuilderData.originalData,
-							item.data
-						);
-						break;
-					case 'section-header':
-						diff = InfoboxBuilderModel.createSectionHeaderDiff(
-							item.infoboxBuilderData.originalData,
-							item.data,
-							item.collapsible
-						);
-						break;
-					default:
-						break;
-				}
-
-				diffs.push(...diff);
-			});
-
-		return diffs;
-	},
-
-	/**
 	 * @desc saves infobox state to MW template
 	 * @returns {Ember.RSVP.Promise}
 	 */
@@ -471,89 +427,6 @@ InfoboxBuilderModel.reopenClass({
 		}
 
 		return item;
-	},
-
-	/**
-	 * @desc Provides information about whether and what elements'
-	 * values has been changed, basing on previously saved start data.
-	 *
-	 * @param {Object} originalData data in form as it was when opening infobox builder
-	 * @param {Object} data current infobox builder state
-	 * @returns {Object}
-	 */
-	createTitleDiff(originalData, data) {
-		const changes = [];
-
-		if (typeof originalData.defaultValue === 'undefined') {
-			return changes;
-		}
-
-		if (data.defaultValue !== originalData.defaultValue) {
-			changes.push({
-				type: 'title',
-				changedField: 'defaultValue'
-			});
-		}
-
-		return changes;
-	},
-
-	/**
-	 * @desc Provides information about whether and what elements'
-	 * values has been changed, basing on previously saved start data.
-	 *
-	 * @param {Object} originalData data in form as it was when opening infobox builder
-	 * @param {Object} data current infobox builder state
-	 * @returns {Object}
-	 */
-	createRowDiff(originalData, data) {
-		const changes = [];
-
-		if (typeof originalData.label === 'undefined') {
-			return changes;
-		}
-
-		if (data.label !== originalData.label) {
-			changes.push({
-				type: 'row',
-				changedField: 'label'
-			});
-		}
-
-		return changes;
-	},
-
-	/**
-	 * @desc Provides information about whether and what elements'
-	 * values has been changed, basing on previously saved start data.
-	 *
-	 * @param {Object} originalData data in form as it was when opening infobox builder
-	 * @param {string} value section header value
-	 * @param {boolean} collapsible
-	 * @returns {Object}
-	 */
-	createSectionHeaderDiff(originalData, value, collapsible) {
-		const changes = [];
-
-		if (typeof originalData.collapsible === 'undefined' || typeof originalData.value === 'undefined') {
-			return changes;
-		}
-
-		if (value !== originalData.value) {
-			changes.push({
-				type: 'section-header',
-				changedField: 'value'
-			});
-		}
-
-		if (collapsible !== originalData.collapsible) {
-			changes.push({
-				type: 'section-header',
-				changedField: 'collapsible'
-			});
-		}
-
-		return changes;
 	}
 });
