+++ front/main/tests/unit/models/infobox-builder-test.js
@@ -776,317 +776,3 @@ test('set edit item', (assert) => {
 		assert.deepEqual(testCase.item.infoboxBuilderData.originalData, testCase.expectedoriginalData);
 	});
 });
-
-test('edit section header item', (assert) => {
-	const originalDataMock = {
-			some_value: 'test'
-		},
-		titleDiff = [
-			{
-				type: 'title',
-				changedField: 'abc'
-			}
-		],
-		rowDiff = [
-			{
-				type: 'row',
-				changedField: 'abc'
-			}
-		],
-		sectionHeaderDiff = [
-			{
-				type: 'section-header',
-				changedField: 'abc'
-			},
-			{
-				type: 'section-header',
-				changedField: 'xyz'
-			}
-		],
-		cases = [
-			{
-				infoboxState: [],
-				expectedDataDiff: []
-			},
-			{
-				infoboxState: [
-					{infoboxBuilderData: {}},
-					{infoboxBuilderData: {}},
-					{infoboxBuilderData: {}}
-				],
-				expectedDataDiff: []
-			},
-			{
-				infoboxState: [
-					{
-						infoboxBuilderData: {},
-						type: 'title'
-					},
-					{
-						infoboxBuilderData: {},
-						type: 'title'
-					},
-					{
-						infoboxBuilderData: {
-							originalData: originalDataMock
-						},
-						type: 'image'
-					}
-				],
-				expectedDataDiff: []
-			},
-			{
-				infoboxState: [
-					{
-						infoboxBuilderData: {
-							originalData: originalDataMock
-						},
-						type: 'title'
-					},
-					{infoboxBuilderData: {}},
-					{infoboxBuilderData: {}}
-				],
-				expectedDataDiff: titleDiff
-			},
-			{
-				infoboxState: [
-					{
-						infoboxBuilderData: {
-							originalData: originalDataMock
-						},
-						type: 'title'
-					},
-					{
-						infoboxBuilderData: {
-							originalData: originalDataMock
-						},
-						type: 'row'
-					},
-					{infoboxBuilderData: {}}
-				],
-				expectedDataDiff: titleDiff.concat(rowDiff)
-			},
-			{
-				infoboxState: [
-					{
-						infoboxBuilderData: {
-							originalData: originalDataMock
-						},
-						type: 'title'
-					},
-					{
-						infoboxBuilderData: {
-							originalData: originalDataMock
-						},
-						type: 'section-header'
-					},
-					{infoboxBuilderData: {}}
-				],
-				expectedDataDiff: titleDiff.concat(sectionHeaderDiff[0], sectionHeaderDiff[1])
-			},
-			{
-				infoboxState: [
-					{
-						infoboxBuilderData: {
-							originalData: null
-						},
-						type: 'title'
-					},
-					{
-						infoboxBuilderData: {
-							originalData: originalDataMock
-						},
-						type: 'row'
-					},
-					{infoboxBuilderData: {}}
-				],
-				expectedDataDiff: rowDiff
-			}
-		];
-
-	sinon.stub(infoboxBuilderModelClass, 'createTitleDiff').returns(titleDiff);
-	sinon.stub(infoboxBuilderModelClass, 'createRowDiff').returns(rowDiff);
-	sinon.stub(infoboxBuilderModelClass, 'createSectionHeaderDiff').returns(sectionHeaderDiff);
-
-	cases.forEach((testCase) => {
-		const model = infoboxBuilderModelClass.create();
-
-		model.set('infoboxState', testCase.infoboxState);
-
-		assert.deepEqual(model.createDataDiffs(), testCase.expectedDataDiff);
-
-	});
-
-	infoboxBuilderModelClass.createTitleDiff.restore();
-	infoboxBuilderModelClass.createRowDiff.restore();
-	infoboxBuilderModelClass.createSectionHeaderDiff.restore();
-});
-
-test('create row diff', (assert) => {
-	const cases = [
-		{
-			originalData: {},
-			data: {},
-			expected: []
-		},
-		{
-			originalData: {},
-			data: {label: 'label'},
-			expected: []
-		},
-		{
-			originalData: {label: 'label'},
-			data: {label: 'label'},
-			expected: []
-		},
-		{
-			originalData: {label: 'label'},
-			data: {},
-			expected: [{
-				type: 'row',
-				changedField: 'label'
-			}]
-		},
-		{
-			originalData: {label: 'test label'},
-			data: {label: ''},
-			expected: [{
-				type: 'row',
-				changedField: 'label'
-			}]
-		},
-		{
-			originalData: {label: ''},
-			data: {label: 'test label'},
-			expected: [{
-				type: 'row',
-				changedField: 'label'
-			}]
-		}
-	];
-
-	cases.forEach((testCase) => {
-		const diff = infoboxBuilderModelClass.createRowDiff(testCase.originalData, testCase.data);
-
-		assert.deepEqual(diff, testCase.expected);
-	});
-});
-
-test('create title diff', (assert) => {
-	const cases = [
-		{
-			originalData: {},
-			data: {},
-			expected: []
-		},
-		{
-			originalData: {},
-			data: {defaultValue: '{{PAGENAME}}'},
-			expected: []
-		},
-		{
-			originalData: {defaultValue: '{{PAGENAME}}'},
-			data: {defaultValue: '{{PAGENAME}}'},
-			expected: []
-		},
-		{
-			originalData: {defaultValue: ''},
-			data: {defaultValue: ''},
-			expected: []
-		},
-		{
-			originalData: {defaultValue: '{{PAGENAME}}'},
-			data: {},
-			expected: [{
-				type: 'title',
-				changedField: 'defaultValue'
-			}]
-		},
-		{
-			originalData: {defaultValue: ''},
-			data: {defaultValue: '{{PAGENAME}}'},
-			expected: [{
-				type: 'title',
-				changedField: 'defaultValue'
-			}]
-		}
-	];
-
-	cases.forEach((testCase) => {
-		const diff = infoboxBuilderModelClass.createTitleDiff(testCase.originalData, testCase.data);
-
-		assert.deepEqual(diff, testCase.expected);
-	});
-});
-
-test('create section-header diff', (assert) => {
-	const cases = [
-		{
-			originalData: {},
-			value: 'some value',
-			collapsible: true,
-			expected: []
-		},
-		{
-			originalData: {
-				value: 'some value',
-				collapsible: true
-			},
-			value: 'some value',
-			collapsible: true,
-			expected: []
-		},
-		{
-			originalData: {
-				value: 'Header',
-				collapsible: false
-			},
-			value: {},
-			collapsible: false,
-			expected: [{
-				type: 'section-header',
-				changedField: 'value'
-			}]
-		},
-		{
-			originalData: {
-				value: 'Header',
-				collapsible: true
-			},
-			value: 'Header',
-			collapsible: false,
-			expected: [{
-				type: 'section-header',
-				changedField: 'collapsible'
-			}]
-		},
-		{
-			originalData: {
-				value: 'Header',
-				collapsible: false
-			},
-			value: 'Header 2',
-			collapsible: true,
-			expected: [
-				{
-					type: 'section-header',
-					changedField: 'value'
-				},
-				{
-					type: 'section-header',
-					changedField: 'collapsible'
-				}
-			]
-		}
-	];
-
-	cases.forEach((testCase) => {
-		const diff = infoboxBuilderModelClass.createSectionHeaderDiff(
-			testCase.originalData,
-			testCase.value,
-			testCase.collapsible
-		);
-
-		assert.deepEqual(diff, testCase.expected);
-	});
-});
